+++ spinosarc_app/__init__.py
@@ -1,3 +1,3 @@
 """SpinoSarc Analyzer - Paraspinal muscle quantification from lumbar MR."""
-__version__ = '0.1.0'
+__version__ = "0.3.0"
 from .analyzer import SpinoSarcAnalyzer, Demographics